add a json parser for input file
add Input and Review classes

added a json parser to parse the input txt file into
an iterator of inputs objects.

To use just create a new instacne of JsonParser.
JsonParser has a input iterator field, which
returns the next json input in the txt file,
on call to getNextInput().

see exapmles on how to use are in JsonParser.main().